Minor changes to CSS+ implementation of HTML
Home Screen.html would be the landing page for the application. This is just a basic copy-paste of the registration for now.

To get to the Registration form from the landing page, until we get a better way for connection; you must change the ... in "action='file:///C: ...;" to the location of the Registration page for it to function on your machine.
Just please don't push any changes here until you change that back (or make your own branch for testing and understanding purposes).

If a new user is detected through a JS comparison between the CWID and DB CWID, then they are brought to the Register New User.html, which includes extra parameters. This is still a skeleton, but is the ideal shape of the registration area.

The CSS file simply contains additional constraints and formatting for the html.